make projects more easily convertable to newer VS versions
git-svn-id: https://svn.code.sf.net/p/pinmame/code/trunk@4060 6f51dfdd-75df-4273-8f3b-818b1e6aa8bf

+++ ext/dmddevice/pinDMDv1/dmddevice_dll/dmddevice.cpp
@@ -40,7 +40,7 @@ DMDDEV int Open()
 	isOpen = false;
 
 	// ftdi devices found
-	UINT8 deviceId = 0;
+	int deviceId = 0;
 	if (numDevs > 0) {
 		// allocate storage for list based on numDevs
 		FT_DEVICE_LIST_INFO_NODE *devInfo = (FT_DEVICE_LIST_INFO_NODE*)malloc(sizeof(FT_DEVICE_LIST_INFO_NODE)*numDevs);
